+++ src/Sequence/Implementation.php
@@ -344,4 +344,11 @@ interface Implementation extends \Countable
      * @return self<T>
      */
     public function dropWhile(callable $condition): self;
+
+    /**
+     * @param callable(T): bool $condition
+     *
+     * @return self<T>
+     */
+    public function takeWhile(callable $condition): self;
 }